Error: Could not find mediaLib accelerator wrapper classes. Continuing in pure Java mode. Occurs in: com.sun.media.jai.mlib.MediaLibAccessor java.lang.NoClassDefFoundError: com/sun/medialib/mlib/Image
Bug is fixed

+++ src/main/java/exercise2/RoiFilter.java
@@ -35,8 +35,8 @@ public class RoiFilter extends DataTransformationFilter2<PlanarImage, PlanarImag
     protected PlanarImage process(PlanarImage entity) {
         PlanarImage image = entity;
         image = PlanarImage.wrapRenderedImage((RenderedImage)image.getAsBufferedImage(rectangle, image.getColorModel()));
-        image.setProperty( "offsetX", rectangle.getX());
-        image.setProperty( "offsetY", rectangle.getY());
+        image.setProperty( "offsetX", (int)rectangle.getX());
+        image.setProperty( "offsetY",(int) rectangle.getY());
         return image;
     }
